rebalance without samples removal

+++ scripts/renameimg.sh
@@ -12,10 +12,19 @@
 # done
 
 # zero filling
+mogrify -format png *.jpg; rm *.jpg;
+
 i=$1
 for file in *.png
 do
         j=$( printf "%d" "$i" )
-        mv "$file" "$j.png"
+        if [ -f "$j.png" ]
+        then
+          echo "File $j.png exists"
+        else
+          echo "mv $file $j.png"
+          mv "$file" "$j.png"
+        fi
+        convert "$j.png" -resize 256x256 "$j.png"
         i=$((i + 1))
 done
